+++ ErsatzTV.Application/Streaming/Queries/GetConcatProcessByChannelNumberHandler.cs
@@ -25,6 +25,7 @@ public class GetConcatProcessByChannelNumberHandler : FFmpegProcessHandler<GetCo
         GetConcatProcessByChannelNumber request,
         Channel channel,
         string ffmpegPath,
+        string ffprobePath,
         CancellationToken cancellationToken)
     {
         bool saveReports = await dbContext.ConfigElements