+++ ErsatzTV.Core/FFmpeg/FFmpegProcessService.cs
@@ -1,10 +1,12 @@
 ﻿using System.Diagnostics;
 using Bugsnag;
 using CliWrap;
+using CliWrap.Buffered;
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.Interfaces.FFmpeg;
 using ErsatzTV.Core.Interfaces.Images;
 using ErsatzTV.FFmpeg.State;
+using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging;
 
 namespace ErsatzTV.Core.FFmpeg;
@@ -15,6 +17,7 @@ public class FFmpegProcessService
     private readonly IImageCache _imageCache;
     private readonly ITempFilePool _tempFilePool;
     private readonly IClient _client;
+    private readonly IMemoryCache _memoryCache;
     private readonly ILogger<FFmpegProcessService> _logger;
     private readonly FFmpegPlaybackSettingsCalculator _playbackSettingsCalculator;
 
@@ -24,6 +27,7 @@ public class FFmpegProcessService
         IImageCache imageCache,
         ITempFilePool tempFilePool,
         IClient client,
+        IMemoryCache memoryCache,
         ILogger<FFmpegProcessService> logger)
     {
         _playbackSettingsCalculator = ffmpegPlaybackSettingsService;
@@ -31,6 +35,7 @@ public class FFmpegProcessService
         _imageCache = imageCache;
         _tempFilePool = tempFilePool;
         _client = client;
+        _memoryCache = memoryCache;
         _logger = logger;
     }
 
@@ -156,6 +161,7 @@ public class FFmpegProcessService
 
     public async Task<Either<BaseError, string>> GenerateSongImage(
         string ffmpegPath,
+        string ffprobePath,
         Option<string> subtitleFile,
         Channel channel,
         Option<ChannelWatermark> globalWatermark,
@@ -190,7 +196,13 @@ public class FFmpegProcessService
                     : None;
 
             Option<WatermarkOptions> watermarkOptions =
-                await GetWatermarkOptions(channel, globalWatermark, videoVersion, watermarkOverride, watermarkPath);
+                await GetWatermarkOptions(
+                    ffprobePath,
+                    channel,
+                    globalWatermark,
+                    videoVersion,
+                    watermarkOverride,
+                    watermarkPath);
 
             FFmpegPlaybackSettings playbackSettings =
                 _playbackSettingsCalculator.CalculateErrorSettings(channel.FFmpegProfile);
@@ -259,6 +271,7 @@ public class FFmpegProcessService
         displaySize.Width != target.Width || displaySize.Height != target.Height;
 
     internal async Task<WatermarkOptions> GetWatermarkOptions(
+        string ffprobePath,
         Channel channel,
         Option<ChannelWatermark> globalWatermark,
         MediaVersion videoVersion,
@@ -290,7 +303,7 @@ public class FFmpegProcessService
                             await watermarkOverride.IfNoneAsync(channel.Watermark),
                             customPath,
                             None,
-                            await _imageCache.IsAnimated(customPath));
+                            await IsAnimated(ffprobePath, customPath));
                     case ChannelWatermarkImageSource.ChannelLogo:
                         Option<string> maybeChannelPath = channel.Artwork
                             .Filter(a => a.ArtworkKind == ArtworkKind.Logo)
@@ -301,7 +314,7 @@ public class FFmpegProcessService
                             maybeChannelPath,
                             None,
                             await maybeChannelPath.Match(
-                                p => _imageCache.IsAnimated(p),
+                                p => IsAnimated(ffprobePath, p),
                                 () => Task.FromResult(false)));
                     default:
                         throw new NotSupportedException("Unsupported watermark image source");
@@ -322,7 +335,7 @@ public class FFmpegProcessService
                             await watermarkOverride.IfNoneAsync(watermark),
                             customPath,
                             None,
-                            await _imageCache.IsAnimated(customPath));
+                            await IsAnimated(ffprobePath, customPath));
                     case ChannelWatermarkImageSource.ChannelLogo:
                         Option<string> maybeChannelPath = channel.Artwork
                             .Filter(a => a.ArtworkKind == ArtworkKind.Logo)
@@ -333,7 +346,7 @@ public class FFmpegProcessService
                             maybeChannelPath,
                             None,
                             await maybeChannelPath.Match(
-                                p => _imageCache.IsAnimated(p),
+                                p => IsAnimated(ffprobePath, p),
                                 () => Task.FromResult(false)));
                     default:
                         throw new NotSupportedException("Unsupported watermark image source");
@@ -342,5 +355,56 @@ public class FFmpegProcessService
         }
 
         return new WatermarkOptions(None, None, None, false);
+    }
+
+    private async Task<bool> IsAnimated(string ffprobePath, string path)
+    {
+        try
+        {
+            var cacheKey = $"image.animated.{Path.GetFileName(path)}";
+            if (_memoryCache.TryGetValue(cacheKey, out bool animated))
+            {
+                return animated;
+            }
+
+            BufferedCommandResult result = await Cli.Wrap(ffprobePath)
+                .WithArguments(
+                    new[]
+                    {
+                        "-loglevel", "error",
+                        "-select_streams", "v:0",
+                        "-count_frames",
+                        "-show_entries", "stream=nb_read_frames",
+                        "-print_format", "csv",
+                        path
+                    })
+                .WithValidation(CommandResultValidation.None)
+                .ExecuteBufferedAsync();
+
+            if (result.ExitCode == 0)
+            {
+                string output = result.StandardOutput;
+                output = output.Replace("stream,", string.Empty);
+                if (int.TryParse(output, out int frameCount))
+                {
+                    bool isAnimated = frameCount > 1;
+                    _memoryCache.Set(cacheKey, isAnimated, TimeSpan.FromDays(1));
+                    return isAnimated;
+                }
+            }
+            else
+            {
+                _logger.LogWarning(
+                    "Error checking frame count for file {File}l exit code {ExitCode}",
+                    path,
+                    result.ExitCode);
+            }
+        }
+        catch (Exception ex)
+        {
+            _logger.LogWarning(ex, "Error checking frame count for file {File}", path);
+        }
+
+        return false;
     }
 }