+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `Reset` - this mode will rebuild the entire playout and will NOT maintain progress
     - This mode is only used when the `Reset Playout` button is clicked on the Playouts page
 - Use ffmpeg to resize images; this should help reduce ErsatzTV's memory use
+- Use ffprobe to check for animated logos and watermarks; this should help reduce ErsatzTV's memory use
 
 ## [0.4.5-alpha] - 2022-03-29
 ### Fixed